'''SONAR''' Fixed some improper memcpy calls, increased bit depth for 10-bit ADCs, added test case for FastSlidingDFT

+++ sandbox/sonar/sonarController/include/SampleDelay.h
@@ -36,7 +36,7 @@ public:
 private:
 	int buflen;
 	adcdata_t *buf, *bufptr, *bufend;
-	int increment;
+	size_t increment;
 };
